# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/users": {
      "get": {
        "description": "Supervisors only see promoters and clients; admins see everyone. New accounts are created by invite text message; if that message cannot be sent, the new user is not saved.",
        "operationId": "AdminUser_listUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of users"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Requires supervisor or admin role"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List users (supervisor / admin)",
        "tags": [
          "Admin users (supervisor / admin)"
        ]
      },
      "post": {
        "description": "Creates a sign-in account and sends instructions by text message. The account is only kept if that message is delivered successfully.",
        "operationId": "AdminUser_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fullName",
                  "phone",
                  "role"
                ],
                "properties": {
                  "fullName": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string",
                    "example": "0244123456"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Optional. Needed on supervisor/admin for Resend alerts."
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "promoter",
                      "client",
                      "supervisor",
                      "admin"
                    ]
                  },
                  "regionId": {
                    "type": "string",
                    "description": "Optional region cuid"
                  },
                  "gender": {
                    "type": "string",
                    "enum": [
                      "male",
                      "female",
                      "other"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          },
          "400": {
            "description": "Validation error, missing SMS configuration, or SMS rejected by provider"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Requires supervisor or admin; only admins may create supervisor/admin"
          },
          "409": {
            "description": "Phone already in use"
          },
          "503": {
            "description": "SMS provider unreachable or HTTP error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create user (invite)",
        "tags": [
          "Admin users (supervisor / admin)"
        ]
      }
    },
    "/admin/users/{id}": {
      "patch": {
        "description": "Partial update. Supervisors may edit promoters and clients only. Only admins may assign supervisor or admin roles.",
        "operationId": "AdminUser_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User id (cuid)",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "nullable": true,
                    "description": "Email for ops alerts / digests, or null to clear"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "promoter",
                      "client",
                      "supervisor",
                      "admin"
                    ]
                  },
                  "regionId": {
                    "type": "string",
                    "nullable": true,
                    "description": "Region id (cuid), or null to remove assignment",
                    "minLength": 1,
                    "maxLength": 64
                  },
                  "isActive": {
                    "type": "boolean"
                  },
                  "gender": {
                    "type": "string",
                    "enum": [
                      "male",
                      "female",
                      "other"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Not allowed for this role or target user"
          },
          "404": {
            "description": "User id not found"
          },
          "409": {
            "description": "Unique constraint violation"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user",
        "tags": [
          "Admin users (supervisor / admin)"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nestlé Ghana Vendor Onboarding API",
    "description": "API documentation for Nestlé Ghana vendor onboarding",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}